requirements.txt
Flask==3.0.0
Pillow==10.4.0
pytest==8.3.4
requests==2.32.3
